pyproject.toml
# note: this file is for local development, currently the file is not used to build
# the plugin

[tool.poetry]
name = "basic-games"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psutil = "5.8.0"
vdf = "3.4"
lzokay = "1.1.5"

[tool.poetry.group.dev.dependencies]
mobase-stubs = { version = "^2.5.0.dev14", allow-prereleases = true }
pyqt6 = "^6.5.2"
pyright = "^1.1.327"
flake8 = "^6.1.0"
flake8-pyproject = "^1.2.3"
ruff = "^0.0.291"
mypy = "^1.5.1"
isort = "^5.12.0"
black = "^23.9.1"
flake8-black = "^0.3.6"
types-psutil = "^5.9.5.16"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]

[tool.black]
exclude = "lib"

[tool.isort]
profile = "black"
multi_line_output = 3
skip = "lib"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
exclude = ["lib", "games/quarantine"]
platform = "win32"

[[tool.mypy.overrides]]
module = "vdf.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lzokay.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pyright]
exclude = ["lib"]
typeCheckingMode = "strict"
pythonPlatform = "Windows"
reportMissingModuleSource = false
